ecommerce-supabase-stripe/requirements.txt
alembic==1.14.0
faiss_cpu==1.10.0
Faker==37.1.0
fastapi==0.115.12
langchain==0.3.23
langchain_community==0.3.21
langchain_core==0.3.54
langchain_google_genai==2.1.3
pydantic==2.11.3
pydantic_settings==2.9.1
SQLAlchemy==2.0.29
sqlmodel==0.0.24
stripe==8.10.0
supabase==2.15.0
uvicorn==0.34.2
